Added Admin Panel

+++ backend/controllers/foodController.js
@@ -36,7 +36,9 @@ const listFood = async (req, res) => {
 const removeFood = async (req, res) => {
     try {
         const food = await foodModel.findById(req.body.id);
-        fs.unlink(`uploads/${food.image}`, ()=>{});
+        if (food.image) {
+            fs.unlink(`uploads/${food.image}`, ()=>{});
+        }
 
         await foodModel.findByIdAndDelete(req.body.id);
         res.json({success:true, message:"Food Removed"});
